+++ CHANGELOG.md
@@ -110,6 +110,7 @@ Please mark all change in change log and use the issue from GitHub
 -   \#1524 Update config "preload_table" description
 -   \#1537 Optimize raw vector and uids read/write
 -   \#1544 Update resources name in HTTP module
+-   \#1546 Move Config.cpp to config directory
 -   \#1567 Update yaml config description
 
 ## Task